提高精度

+++ engine/src/main/java/com/sunshine/engine/base/Entity.java
@@ -124,13 +124,13 @@ public abstract class Entity {
 
   public void mergeDrawInfo() {
     // move
-    drawInfo.rcDst.left = (int) (drawInfo.rcSrc.left * scale) + drawArea.l;
-    drawInfo.rcDst.top = (int) (drawInfo.rcSrc.top * scale) + drawArea.t;
-    drawInfo.rcDst.right = (int) (drawInfo.rcSrc.right * scale) + drawArea.l;
-    drawInfo.rcDst.bottom = (int) (drawInfo.rcSrc.bottom * scale) + drawArea.t;
+    drawInfo.rcDst.left = drawInfo.rcSrc.left * scale + drawArea.l;
+    drawInfo.rcDst.top = drawInfo.rcSrc.top * scale + drawArea.t;
+    drawInfo.rcDst.right = drawInfo.rcSrc.right * scale + drawArea.l;
+    drawInfo.rcDst.bottom = drawInfo.rcSrc.bottom * scale + drawArea.t;
     // rotate
-    drawInfo.ptDst.x = (int) (drawInfo.ptSrc.x * scale) + drawArea.l;
-    drawInfo.ptDst.y = (int) (drawInfo.ptSrc.y * scale) + drawArea.t;
+    drawInfo.ptDst.x = drawInfo.ptSrc.x * scale + drawArea.l;
+    drawInfo.ptDst.y = drawInfo.ptSrc.y * scale + drawArea.t;
   }
 
   public void isMute(boolean mute) {
